added the functionality of proceesing declarations ate every method entry/exit
In this change Chicory now is constructing declaration records for all classes any time a method is entered or exited. Also, it adds list ellements to the ListInfo node if they are non-primitive throgh the newly defined variable ListElement. The declaration writing is now behaving correclty. However, hte traces writing for the new variable through returning null when calling getValFromParent(). This is to be fixed in the new change.