Applied an enhancment to the performance issue
Now declarations are only processed for methods that we are about to write traces for. This new version whenever it attempt to write traces for a method (whether for entry or exit), enhanced Chciory will look at the method and construct declaration trees for both entry and exit of the method. It is possible to change this to be focused only on the entry or the exit of the method to have even better perofrmance. However, the change needs to be palced is too much for the gain we are to get. Thus, I left this minor step unenhanced.