enahnced Chicory is Fully optimized
I found out that its is not as diffcult as it seems to make the enhnaced Chicory construct decl trees for method entry or exit as needed. So, I made the change test the resutlt given by the unifier or the Daikon itself ar the same. Thus, the change is correct.