added hocks to be able to identify packages, classes and method in the ppt and traces
This is the first step toward combining objects. Bfore proceeding, I need to identify what needs to be merged when combining two objects and what is not. Also, there must be a methoid naming convetion the I should be following on the fly (as the traces are written) because it is expensive to create new objcts (ppts and traces).