Introduced combining objects ability
As of the front-end (my unifier front-end) there are no errors. However, it is highly expected that Daikon will not be happy with the ppts and traces it reads. One of the things I'm expecting it to yell about is the nonce values. I don't change those value as of this commit to sustain integrity as much as possible. But I don't think that Daikon will be happy of finding two different methods from two different objects with the same nonce.

Also, there could be problems arising from the properties of each variable in the ppt definition. I didn't not want to considering changing it because it is a lot of work. If Daikon doesn't complain about it that would be great. However, if it complains, I would need to read properties carefully (as Daikon would do) and change them according to my combining strategy.