Fixed all the issue in the combination technique I introduced.
Daikon now reads and generates invariants based on the combined objects methodology successfully. I don't think that there will be an issue in feeding those results to TEMI to generate an FSM and analyzing it, but this should be tested. If any issue arise because of TEMI it might be too complicated to fix, thus it is better to analysis the given inv in that case rather than fixing the given issue.