Pushed changes that was found on the PptInfo.java and released a new JAR.
The new JAR is just to make sure what version I'm working on in the different
 machines I access this repo with.

The PptInfo.java found with some changes that are not commited. The changes a
re reflected in the JAR file. So, I thought I might as well push the chnages
of that file.

The changes on the .gitignore file are made to make sure we ignore the Daikon
's or Chicory's generated files.